Create nullDebuggerWriter
I created doc.go to contain the documentation for the debugger package.

I created debugger.go and implemented nullDebuggerWriter.
nullDebuggerWriter is the default implementation of debugger.Writer used
on platforms that do not support outputting messages to a debugger.

I created debugger_test.go and implemented the unit tests for
nullDebuggerWriter.